+++ src/irx/analysis/handlers/__init__.py
@@ -1,0 +1,6 @@
+"""
+title: Semantic-analysis handlers.
+summary: >-
+  Group semantic node handlers and shared visitor mixins by concern while
+  keeping the public analyzer class small and focused on orchestration.
+"""
