+++ src/irx/analysis/handlers/declarations.py
@@ -10,11 +10,11 @@ summary: >-
 from __future__ import annotations
 
 from irx import astx
-from irx.analysis.validation import validate_assignment
-from irx.analysis.visitors.base import (
+from irx.analysis.handlers.base import (
     SemanticAnalyzerCore,
     SemanticVisitorMixinBase,
 )
+from irx.analysis.validation import validate_assignment
 from irx.typecheck import typechecked
