+++ src/irx/analysis/handlers/imports.py
@@ -10,7 +10,7 @@ summary: >-
 from __future__ import annotations
 
 from irx import astx
-from irx.analysis.visitors.base import (
+from irx.analysis.handlers.base import (
     SemanticAnalyzerCore,
     SemanticVisitorMixinBase,
 )
