--- src/irx/analysis/visitors/__init__.py
@@ -1,6 +1,0 @@
-"""
-title: Semantic-analysis visitor mixins.
-summary: >-
-  Group semantic node handlers by concern while keeping the public analyzer
-  class small and focused on orchestration.
-"""
